phpstan lvl 1

+++ inc/replace.class.php
@@ -165,7 +165,7 @@ class PluginUninstallReplace extends CommonDBTM
                 ]);
             }
 
-           // General Informations - NAME
+           // General Information - NAME
             if ($model->fields["replace_name"]) {
                 if ($overwrite || empty($newitem->fields['name'])) {
                     $newitem->update(
@@ -465,8 +465,8 @@ class PluginUninstallReplace extends CommonDBTM
 
                    // If old item is attached in PDF/CSV
                    // Delete AND Purge it in DB
-                    if ($document_added) {
-                           $olditem->delete(['id' => $olditem_id], true);
+                    if (isset($document_added) && $document_added) {
+                       $olditem->delete(['id' => $olditem_id], true);
                     }
                     break;
